Move contents of many plugin headers to corresponding source files in TopJetCombination
List of classes:
- TtSemiLepJetCombGeom
- TtSemiLepJetCombMaxSumPtWMass
- TtSemiLepJetCombWMassDeltaTopMass
- TtSemiLepJetCombWMassMaxSumPt
- TtSemiLepHypothesis-derived classes
- TtFullLepHypothesis-derived classes
- TtFullHadHypothesis-derived classes
