+++ TopQuarkAnalysis/TopJetCombination/plugins/TtFullHadHypGenMatch.cc
@@ -1,5 +1,24 @@
-#include "TopQuarkAnalysis/TopJetCombination/plugins/TtFullHadHypGenMatch.h"
+#include "TopQuarkAnalysis/TopJetCombination/interface/TtFullHadHypothesis.h"
 #include "AnalysisDataFormats/TopObjects/interface/TtFullHadEvtPartons.h"
+#include "DataFormats/Candidate/interface/LeafCandidate.h"
+
+class TtFullHadHypGenMatch : public TtFullHadHypothesis {
+public:
+  explicit TtFullHadHypGenMatch(const edm::ParameterSet& cfg);
+  ~TtFullHadHypGenMatch() override;
+
+private:
+  /// build the event hypothesis key
+  void buildKey() override { key_ = TtFullHadronicEvent::kGenMatch; };
+  /// build event hypothesis from the reco objects of a semi-leptonic event
+  void buildHypo(edm::Event& evt,
+                 const edm::Handle<std::vector<pat::Jet> >& jets,
+                 std::vector<int>& match,
+                 const unsigned int iComb) override;
+
+protected:
+  edm::EDGetTokenT<TtGenEvent> genEvtToken_;
+};
 
 TtFullHadHypGenMatch::TtFullHadHypGenMatch(const edm::ParameterSet& cfg)
     : TtFullHadHypothesis(cfg), genEvtToken_(consumes<TtGenEvent>(edm::InputTag("genEvt"))) {}
@@ -56,3 +75,6 @@ void TtFullHadHypGenMatch::buildHypo(edm::Event& evt,
     }
   }
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(TtFullHadHypGenMatch);
