+++ TopQuarkAnalysis/TopJetCombination/plugins/TtSemiLepHypGeom.cc
@@ -1,16 +1,13 @@
-#ifndef TtSemiLepHypMaxSumPtWMass_h
-#define TtSemiLepHypMaxSumPtWMass_h
-
 #include "TopQuarkAnalysis/TopJetCombination/interface/TtSemiLepHypothesis.h"
 
-class TtSemiLepHypMaxSumPtWMass : public TtSemiLepHypothesis {
+class TtSemiLepHypGeom : public TtSemiLepHypothesis {
 public:
-  explicit TtSemiLepHypMaxSumPtWMass(const edm::ParameterSet& cfg) : TtSemiLepHypothesis(cfg){};
-  ~TtSemiLepHypMaxSumPtWMass() override{};
+  explicit TtSemiLepHypGeom(const edm::ParameterSet& cfg) : TtSemiLepHypothesis(cfg){};
+  ~TtSemiLepHypGeom() override{};
 
 private:
   /// build the event hypothesis key
-  void buildKey() override { key_ = TtSemiLeptonicEvent::kMaxSumPtWMass; };
+  void buildKey() override { key_ = TtSemiLeptonicEvent::kGeom; };
   /// build event hypothesis from the reco objects of a semi-leptonic event
   void buildHypo(edm::Event& evt,
                  const edm::Handle<edm::View<reco::RecoCandidate> >& leps,
@@ -22,4 +19,5 @@ private:
   };
 };
 
-#endif
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(TtSemiLepHypGeom);
